debug.jl

AdalmPluto.libIIO_jl.C_iio_device_get_sample_size

AdalmPluto.libIIO_jl.C_iio_device_identify_filename

AdalmPluto.libIIO_jl.C_iio_device_get_sample_size — Method

C_iio_device_get_sample_size(device)

Get the current sample size.

Parameters

device::Ptr{iio_device} : A pointer to an iio_device structure

Returns

On success, the sample size in bytes

On error, a negative errno code is returned

NOTE

The sample size is not constant and will change when channels get enabled or disabled.

libIIO documentation

AdalmPluto.libIIO_jl.C_iio_device_identify_filename — Method

C_iio_device_identify_filename(device, filename)

Identify the channel or debug attribute corresponding to a filename.

Parameters

device::Ptr{iio_device} : A pointer to an iio_device structure

filename::String : A NULL-terminated string corresponding to the filename

Returns

On success, (0, channel::Ptr{iio_channel}, attribute::String) is returned.

On error, (errno, NULL, "") is returned, where errno is a negative error code.

libIIO documentation